fix(NeqMecuN): fix "too many connections" MySQL error with using tearDown method

+++ test/old/DataStore/CsvBaseTest.php
@@ -30,7 +30,13 @@ class CsvBaseTest extends AbstractTest
 
     protected function tearDown(): void
     {
-        unlink($this->filename);
+        try {
+            if (is_file($this->filename)) {
+                unlink($this->filename);
+            }
+        } finally {
+            parent::tearDown();
+        }
     }
 
     protected function _initObject($data = null)
